Split accountSummary component in two
